+++ packages/components/src/components/all.ts
@@ -1,0 +1,9 @@
+import * as components from './index.js';
+
+/**
+ * List of all components in the library.
+ * Used mainly for quickStart functionality
+ */
+export const ALL_COMPONENTS = [
+  ...Object.values(components),
+];