+++ packages/components/src/components/icon/template.ts
@@ -1,0 +1,8 @@
+import { html } from '@microsoft/fast-element';
+import { Icon } from './class.js';
+
+export const iconTemplate = () => html<Icon>`
+  <span class="icon" part="root">
+    <slot></slot>
+  </span>
+`;