+++ packages/components/src/components/logo/class.ts
@@ -1,6 +1,6 @@
 import { attr } from '@microsoft/fast-element';
 import { FoundationElement } from '@microsoft/fast-foundation';
-import type { LogoVariant } from './options';
+import type { LogoVariant } from './options.js';
 
 /**
  * The logo represents the company and forms the core of the brand identity.
@@ -28,10 +28,4 @@ export class Logo extends FoundationElement {
    */
   @attr
   public variant: LogoVariant = 'color';
-}
-
-declare global {
-  interface HTMLElementTagNameMap {
-    'sds-logo': Logo;
-  }
 }