+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 Please have a look at the corresponding [packages](./packages) for a list of recent changes:
 
 - [@sick-design-system/css](./packages/css/CHANGELOG.md)
+- [@sick-design-system/components](./packages/components/CHANGELOG.md) 
 - [@sick-design-system/design-tokens](./packages/design-tokens/CHANGELOG.md)
 - [@sick-design-system/docs](./packages/docs/CHANGELOG.md)
 - [@sick-design-system/eslint-config-sds](./packages/eslint-config-sds/CHANGELOG.md)